modified Modal component using material Ui
