+++ package.json
@@ -11,9 +11,11 @@
     "@types/node": "^12.19.1",
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
+    "@types/styled-components": "^5.1.4",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
     "react-scripts": "3.4.3",
+    "react-stylesheet": "^2.2.2",
     "typescript": "^3.7.5"
   },
   "scripts": {
